requirements.txt
boto3==1.19.9
google-api-python-client==2.29.0
google-auth==2.3.3
numpy==1.17.3
oauth2client==4.1.3
opencv-python==4.5.4.58
python-dateutil==2.8.2
PyYAML==6.0
rsa==4.0
s3transfer==0.5.0
urllib3==1.26.7
youtube-dl==2021.12.17
youtube3==1.2.3